Refactored Case management component to top level directory and made case management as the default component to be displayed
