+++ client/shell/src/app/case-management/case-management-routing.module.ts
@@ -2,10 +2,10 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 import { LoginGuard } from '../core/auth/_guards/login-guard.service';
-import { CaseManagementComponent } from './case-management/case-management.component';
+import { CaseManagementComponent } from './case-management.component';
 
 const routes = [{
-  path: 'home',
+  path: 'case-management',
   component: CaseManagementComponent,
   canActivate: [LoginGuard]
 }];
