+++ client/shell/src/app/case-management/case-management.component.html
@@ -1,0 +1,4 @@
+<md-tab-group>
+  <md-tab label="My Schools">Content 1</md-tab>
+  <md-tab label="Monthly Report">Content 2</md-tab>
+</md-tab-group>
