+++ client/shell/src/app/case-management/case-management.component.ts
