--- client/shell/src/app/case-management/case-management/case-management.component.html
@@ -1,3 +1,0 @@
-<p>
-  case-management works!
-</p>
